requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
boto3==1.34.14
botocore==1.34.14
click==8.1.7
common-fate-schema==0.7.0
Django==4.2.9
django-cors-headers==4.3.1
django-oauth2==4.1
django-ratelimit==4.1.0
et-xmlfile==1.1.0
graphene==3.3
graphene-django==3.1.5
graphene-federation==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
jmespath==1.0.1
MarkupSafe==2.1.5
openpyxl==3.1.2
promise==2.3
provider==0.11.0
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
s3transfer==0.10.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.9.0
urllib3==1.26.18
pytest==7.4.4
pytest-django==4.7.0
pytest-watch==4.2.0
ruff==0.3.7
django-auditlog==3.0.0
djangorestframework==3.15.2
pytz==2024.1
Jinja2==3.1.3
weasyprint==59.0
djangorestframework-simplejwt==5.3.1
